WebSockets/tests/submissions/Opera/constructor/002.html

<!doctype html>
<title>WebSockets: new WebSocket(invalid url)</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
function checkException(e) {
  assertInstanceOf(e, DOMException);
  assert_equals(e.code, e.SYNTAX_ERR);
}
assertThrows(function(){new WebSocket("/test")}, checkException);
assertThrows(function(){new WebSocket("ws://foo bar.com/")}, checkException);
assertThrows(function(){new WebSocket("wss://foo bar.com/")}, checkException);
assertThrows(function(){new WebSocket("http://"+location.hostname+"/")}, checkException);
assertThrows(function(){new WebSocket("mailto:example@example.org")}, checkException);
assertThrows(function(){new WebSocket("about:blank")}, checkException);
assertThrows(function(){new WebSocket("ws://"+location.hostname+"/#")}, checkException);
assertThrows(function(){new WebSocket("ws://"+location.hostname+"/#test")}, checkException);
assertThrows(function(){new WebSocket("?test")}, checkException);
assertThrows(function(){new WebSocket("#test")}, checkException);
end();
</script>
